tests: run timing test last
The timing decorator modifies global
state which cannot be undone, so run
that test last.
